backend/requirements.txt

flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
stripe==4.2.0
werkzeug==2.2.3
pypdf==3.5.1
python-docx==0.8.11
